Renamed MSOfficeDocumentParser to MsOfficeDocumentParser
